a new test file is created for running the test cases
